Begin splitting the monolith — config & settings data moved to core/config/ .
lora_trainer_gui.py is 27,640 lines in a single file. At that size:

- Navigation and review. The file cannot be read or reviewed whole; every
  lookup means scrolling thousands of unrelated lines.
- Change safety. Any edit risks touching neighbouring code that is not visible
  in the current context; regressions surface only when the app is run.
- Testability. Modules wired to one monolithic self object cannot be isolated
  for unit tests without booting the whole GUI.
- Parallel work. Two developers (or two branches) are near-certain to collide
  in a single file on any non-trivial change.

This branch is a pure structural decomposition (PURE MOVE). Code is relocated
from the monolith into focused modules under fizgig_gui/core/, grouped by
responsibility: config/, domain/, ui_base/, tabs/, models/, training/,
dataset/ and a thin shell.py. Logic, signatures and initialisation order are
untouched — behaviour is bit-for-bit identical to master.

Extraction proceeds leaf-first (config/domain first, then the UI base, the
tabs, the training engine, and finally the app shell) so that each slice shrinks
the monolith and makes the next one shorter and safer.

Important: this branch is separation only. Each extracted module still needs its
own per-module refactoring (decoupling from the monolith self, removing
forward references, unit tests) — that work lands separately, one module at a
time.

The first commit relocates configuration data (constants, preferences, presets,
last-used state, settings map) and minimax_math into core/config/ and
core/domain/, and adds core/paths.py. Monolith: 27,640 → 26,551 lines.
