+++ lora_trainer_gui.py
@@ -51,61 +51,11 @@ except ImportError:
     FaceDetector = None
     FaceEmbedder = None
 
-# Refined Color Palette (Fizgig Visual Style Guide)
-COLORS = {
-    "bg_deep": "#1E2530",        # Main window background
-    "bg_surface": "#252D38",     # Cards, panels, inputs
-    "bg_hover": "#2A3542",       # Hover states
-    "bg_header": "#1A2028",      # Collapsible section headers
-
-    "text_primary": "#F0F4F8",   # Main text
-    "text_secondary": "#8A9BAE", # Labels
-    # Explanatory prose — card descriptions, row hints, fine print. Its own tier because the
-    # muted grey below reads at 2.54:1 on a card, which fails WCAG even for large text, and this
-    # is the copy that actually explains the app. 8.64:1, still a step down from text_primary so
-    # a card title stays visibly louder than its explanation.
-    "text_explain": "#C3CDD9",
-    # Genuinely de-emphasised UI only: disabled controls, placeholders, and the one-word captions
-    # beside widgets ("seed", "W", "H"). NOT for prose — that is text_explain above.
-    "text_muted": "#5A6B7E",
-
-    "accent": "#3B82F6",         # Primary actions, links
-    "accent_hover": "#60A5FA",   # Accent hover
-    "accent_subtle": "#1E3A5F",  # Accent backgrounds
-
-    # Training-queue button (status bar, lower right). Deliberately a LIGHT blue: on a dark
-    # bar a pale block reads as a distinct control, where another dark-surface panel just
-    # blends into the furniture. Text on it is bg_deep (8.6:1) — text_primary would vanish.
-    "queue_blue": "#93C5FD",
-    "queue_blue_hover": "#BFDBFE",
-
-    "border": "#3A4555",         # Borders, dividers
-    "border_focus": "#3B82F6",   # Focus rings
-
-    # Scrollbar thumb — the accent blue, matching the selected tab. Named entries rather than
-    # pointing straight at "accent" so this stays tunable without dragging tabs and links along
-    # with it. 4.19:1 against the trough, up from the 1.06:1 it used to be.
-    "scrollbar_thumb": "#3B82F6",
-    "scrollbar_thumb_hover": "#60A5FA",
-
-    "success": "#10B981",        # Success states
-    "warning": "#F59E0B",        # Warnings
-    "error": "#EF4444",          # Errors
-}
-
-# Typography
-FONT_FAMILY = "Segoe UI"
-FONT_MONO = "Consolas"
-
-# Legacy color constants (for backwards compatibility during transition)
-BG_COLOR = COLORS["bg_deep"]
-FG_COLOR = COLORS["text_primary"]
-ACCENT_COLOR = COLORS["accent"]
-ENTRY_BG = COLORS["bg_surface"]
-BUTTON_ACTIVE = COLORS["bg_hover"]
-BORDER_COLOR = COLORS["border"]
-ACTIVE_ENTRY_BG = "white"  # Background color for active entry field
-ACTIVE_ENTRY_FG = "black"  # Text color for active entry field
+from fizgig_gui.core.config.constants import (
+    COLORS, FONT_FAMILY, FONT_MONO,
+    BG_COLOR, FG_COLOR, ACCENT_COLOR, ENTRY_BG, BUTTON_ACTIVE, BORDER_COLOR,
+    ACTIVE_ENTRY_BG, ACTIVE_ENTRY_FG, SAMPLE_RESOLUTIONS,
+)
 
 
 class _GUIWriter:
@@ -455,1078 +405,39 @@ LAST_TRAIN_FILE = os.path.join(PRESETS_DIR, ".last_train_settings.json")
 # never auto-starts on launch (a queue found at startup waits for the user's first Start).
 QUEUE_FILE = os.path.join(PRESETS_DIR, "training_queue.json")
 
-# MiniMax H3 — how much of training happens at LOW NOISE ("Low-noise training" on the tab).
-#
-# The dial is the SHARE, not a shift number. Same underlying knob, stated as the thing that
-# actually matters. Training draws u uniform and maps it through
-#
-#     sigma = shift * u / (1 + (shift - 1) * u)
-#
-# so the fraction of steps landing below a threshold T is P = T / (shift*(1 - T) + T). At
-# T = 0.5 that inverts to exactly
-#
-#     shift = (1 - P) / P
-#
-# — no solver, no table, no resolution term. Reference points: 50% -> shift 1 (unshifted
-# uniform), 22% -> 3.5, 7.7% -> 12 (H3's own video default).
-#
-# DELIBERATELY UNCAPPED and deliberately independent of megapixels. The previous MP-keyed
-# version derived its numbers from a sqrt-token transfer rule that has never been validated on
-# H3, and capping the list to that rule's answers ruled out settings worth trying.
-MINIMAX_LOWNOISE_SIGMA = 0.5     # "low noise" = the cleaner half of the sigma range
-
-
-def minimax_lownoise_to_shift(pct):
-    """Share of steps below sigma 0.5 (percent) -> the shift that produces it. None if unusable.
-
-    0 and 100 are the asymptotes, not values: 0% needs an infinite shift and 100% a shift of 0,
-    and neither is a schedule. Everything strictly between them is allowed."""
-    try:
-        p = float(str(pct).strip().rstrip("%")) / 100.0
-    except (TypeError, ValueError):
-        return None
-    if not (0.0 < p < 1.0):
-        return None
-    return (1.0 - p) / p
-
-
-def minimax_highnoise_lr(pct):
-    """'Medium to High LR adjustment' (percent) -> a plain multiplier. None if unusable.
-
-    Applies to steps drawn ABOVE sigma 0.5 — the same threshold the low-noise box is defined
-    against, so the two controls always agree about where the boundary is. 100 means unchanged.
-    0 is allowed and means those steps train nothing; the ceiling is 100 because raising the LR
-    for the noisy end is a different experiment from damping it, and one dial should do one thing.
-    """
-    try:
-        p = float(str(pct).strip().rstrip("%")) / 100.0
-    except (TypeError, ValueError):
-        return None
-    return None if not (0.0 <= p <= 1.0) else p
-
-
-# The logit-normal ("mid-concentrated") variant of the above is GONE, deliberately.
-#
-# It bunched training around the middle and thinned BOTH tails, and the tail it was quietly
-# deleting is the high-noise end where pose and composition are decided — at 60% low noise it left
-# 0.7% of a run above sigma 0.9. A 20-step render spends most of its steps there, so the LoRA was
-# being asked to hold structure it had barely trained on: fine under a 4-step Turbo workflow,
-# soft or distorted without one.
-#
-# Tested directly (Peter, 14 Aug): the Likeness preset with mid-concentrated OFF works at strength
-# 1.0 without Turbo, and likeness did not suffer — so it was not buying what it was supposed to buy
-# either. The 60% share was never the problem.
-
-
-def minimax_shift_to_lownoise(shift):
-    """The inverse — shows an existing shift (e.g. from an older preset) as a percentage."""
-    try:
-        s = float(shift)
-    except (TypeError, ValueError):
-        return None
-    return None if s <= 0 else 100.0 / (1.0 + s)
-
-
-# MiniMax H3 — which of the 50 DiT blocks to train ("Blocks to Train" on the Training tab).
-#
-# THIS IS A SEARCH, not a recommendation, and the labels say so. H3 is 50 identical blocks with
-# no published map. The ranges come from the proportions of Klein's empirically-built map
-# (composition in the first ~30% of depth, identity ~30-75%, fine detail ~60-100%) scaled to 50
-# blocks — an analogy between two architectures, not a measurement of this one.
-#
-# READ EVERY RESULT AS: "was this subset sufficient?" A range that trains a good likeness means
-# its COMPLEMENT was not needed — whichever end that turns out to be. There is no option here
-# that disproves the idea; a surprising winner just relocates the answer. (An earlier revision
-# labelled the front half a "control, expected worse". That was wrong: a good front-half result
-# would mean the back half is droppable, which is the outcome we are hunting, not a refutation.)
-#
-# Why dropping blocks may HELP rather than merely cost less: a block that carries no identity
-# still receives gradient, and what is left for it to learn is the dataset's background, framing
-# and lighting. Excluding it removes capacity that would otherwise go into memorising the set.
-#
-# What is measured: per-block ||dW|| on two finished H3 LoRAs is nearly FLAT (3x quietest to
-# loudest against a 2% flat expectation, thirds within 28/34/38), and the quietest blocks do not
-# agree between runs. Weight movement offers no map either way — which is why the only way to
-# learn anything here is to train a range and compare.
-# The box is EDITABLE: these are jumping-off points, and anything the trainer's parser accepts
-# can be typed instead — "3-12, 14-15, 22, 31-33". The label separator is "·" and not "-" or a
-# plain space, because both of those appear inside a block spec; splitting on them would turn
-# "3-12, 14" into "3-12," and train the wrong set without ever complaining.
-# MiniMax H3 — "Training Structure" on the Training tab. One decision, two numbers behind it.
-#
-#   pct  = share of steps trained below sigma 0.5 (the clean end, where detail and identity live).
-#          Converted to the trainer's --shift by minimax_lownoise_to_shift.
-#   lr   = what the steps ABOVE that threshold do to the learning rate, as a percentage.
-#
-# Both presets recommend 100 — the noisy steps train at full rate — because that is the only
-# setting anything has been measured at. The dial exists because dropping the clean-end share
-# makes high-noise steps the MAJORITY (92% at the model's own schedule against 40% here), and the
-# worry was that they would swamp the few clean-end steps carrying identity.
-#
-# Measured across FIVE datasets (Peter, 14 Aug), at BOTH densities and at 0% and 100%: nothing
-# corrupts in any of the four combinations, and 100% has visibly better face SHAPE every time.
-# That fits — shape is settled early, at high noise, while skin and texture come late — so damping
-# costs geometry rather than buying stability, including at 8% clean-end where the noisy end is
-# 92% of the run and the swamping worry should have been strongest. It was not. Nothing is damped
-# by default.
-#
-# The box stays because it is the SAFE way to bias a run toward surface detail. Mid-concentrated
-# tried to do that by changing which noise levels were SAMPLED, which took the adapter
-# off-distribution and distorted at 20 steps. This changes only how much is learned from them —
-# the schedule the model sees is untouched, which is why 0% renders cleanly at either density. A
-# skin-texture LoRA is the obvious use.
-#
-# The 8% is not a guess. ai-toolkit's H3 entry overrides its global 'sigmoid' timestep type with
-# 'shift' (ui/src/app/jobs/new/options.tsx) against a scheduler at the model's RELEASED video flow
-# shift of 12 (minimax_h3.py + packing.py). A shifted-uniform draw at shift 12 puts 1/13 = 7.7% of
-# steps below sigma 0.5. It is the schedule the model's own flow shift implies — not a published
-# statement of what MiniMax ran in pre-training, which is why the label says the former.
-# There is deliberately no "style" setting between these two. Style lives at the CLEAN end, not
-# the noisy one — Fizgig's own Klein work established that, extracting at three timestep ranges
-# and finding style concentrated in the late/clean band. Brushwork, palette and grain are surface
-# properties, so a style LoRA wants the same density a likeness one does; what distinguishes it is
-# often rank and LR, not the noise schedule.
-# An earlier revision of this shipped a "Balanced / style — 25%" option built by treating style as
-# composition and pushing AWAY from the clean end. That was backwards, and a mislabelled setting is
-# worse than a missing one.
-MINIMAX_STRUCTURE_OPTIONS = {
-    "Likeness and Style — 60% clean-end": (60, 100),
-    "Model default, movement — 8% clean-end": (8, 100),
-    "Custom": None,
-}
-MINIMAX_STRUCTURE_DESC = {
-    "Likeness and Style — 60% clean-end":
-        "Most of the run on nearly-clean images. Skin, hair and identity are learned there — and "
-        "so is style, which is a surface property rather than a compositional one. The tuned "
-        "default for stills.",
-    "Model default, movement — 8% clean-end":
-        "The schedule H3's own flow shift implies, and what the reference trainer uses. Weighted "
-        "to movement and composition rather than fine detail.",
-    "Custom":
-        "Type your own share. Below ~50% the high-noise steps become the majority, which is what "
-        "the LR adjustment beside this is for.",
-}
-MINIMAX_STRUCTURE_DEFAULT = "Likeness and Style — 60% clean-end"
-
-MINIMAX_BLOCK_OPTIONS = [
-    "all · every block (50 of 50)",
-    "10-49 · skip the first 10",
-    "14-37 · middle band",
-    "25-49 · back half",
-    "0-24 · front half",
-]
-
-
-MINIMAX_NUM_BLOCKS = 50          # H3's DiT block count (MiniMaxH3Config.num_layers)
-
-# Optimised Likeness Learning — the block set photo steps train when the checkbox is on (clips
-# always train the full model). 20-49 is the measured recipe (19 Aug): the visual-exclusive band
-# 20-26 adds pose/likeness stability, identity lives 27-49, and the front trunk 0-19 is where
-# photo gradients deform anatomy. One place to tweak as the add-back ladder refines the figures.
-MINIMAX_LIKENESS_BLOCKS = "20-49"
-
-# Voice routing — the block set audio-only steps train. 34-49 per the block map (audio core
-# 38-48 peak 41-42, shoulder 34-37) and Peter's A/B (24 Aug): audio-only trained at 34-49 is
-# clean; at 20-49 the audio training corrupted the visual blocks. Clips still train the full
-# model (pending the same test for video).
-MINIMAX_AUDIO_BLOCKS = "34-49"
-
-# Base Precision — the label the user sees, and the --base_quant value it sends. Auto plans the
-# quantisation and the block-swap count together (see plan_base_quant in minimax/trainer.py);
-# an explicit pick is never overridden, the swap plan is built around it instead.
-MINIMAX_BASE_QUANT_OPTIONS = [
-    "Auto (recommended)",
-    "int8 · most accurate, needs ~30 GB free",
-    "4-bit · fits smaller cards",
-]
-
-
-def minimax_base_quant(raw):
-    """Dropdown label -> the --base_quant value. Anything unrecognised falls back to auto."""
-    s = str(raw or "").split("·")[0].strip().lower()
-    if s.startswith("int8"):
-        return "int8"
-    if s.startswith("4-bit") or s.startswith("nf4"):
-        return "nf4"
-    return "auto"
-
-
-def minimax_block_spec(raw):
-    """The block selection out of a dropdown label OR a hand-typed spec. "" -> "all"."""
-    return str(raw or "").split("·")[0].strip() or "all"
-
-
-# Training Base — which H3 fine-tune the LoRA trains against. fl2va (first/last-frame) is the
-# ordinary model; ref2va is the Reference-to-Video fine-tune ComfyUI's r2v workflow runs. Same
-# architecture and shapes, so the trainer takes either — a LoRA is most faithful on the base it
-# trained against, which is the point of the choice. Deliberately NOT preset-affected: the var
-# lives outside self.entries and is never collected, so presets/last-train can't flip it.
-MINIMAX_TRAIN_BASE_OPTIONS = [
-    "First/last frame (fl2va) — standard",
-    "Reference (ref2va)",
-]
-
-
-def minimax_train_base(raw):
-    """Dropdown label -> canonical base key. Anything unrecognised is the fl2va default."""
-    return "ref2va" if "ref2va" in str(raw or "").lower() else "fl2va"
-
-
-def ft_checkpoint_continuation(path):
-    """(next_window, epochs_done) for a rotation fine-tune checkpoint.
-
-    A fine-tune's continuation point is the full checkpoint itself, not a state dir. The
-    trainers stamp `fizgig_next_start_window` (where the rotation cycle picks back up) and
-    `fizgig_ft_epochs_done` (cumulative epochs trained) into every FT save; the epoch count
-    also lives in the filename's -NNNNNN, which wins when present (the un-numbered FINAL
-    artifact has only the metadata). Anything missing reads as 0 — a continuation that
-    restarts the cycle at window 0 is degraded, not wrong."""
-    next_window, epochs_done = 0, 0
-    try:
-        from safetensors import safe_open
-        with safe_open(path, framework="pt") as f:
-            md = f.metadata() or {}
-        next_window = int(md.get("fizgig_next_start_window", 0))
-        epochs_done = int(md.get("fizgig_ft_epochs_done", 0))
-    except Exception:
-        pass
-    m = re.search(r"-(\d{6})\.safetensors$", os.path.basename(path or ""))
-    if m:
-        epochs_done = int(m.group(1))
-    return next_window, epochs_done
-
-
-# Built-in presets — always available in the Load Preset dropdown, prefixed with ✨ to distinguish
-# from user-saved presets. Defined in code so they ship with the app and can't be deleted accidentally.
-# Tune these as empirical findings accumulate.
-BUILT_IN_PRESETS = {
-    "✨ Old Reliable (rank 16, full model, single subject)": {
-        "NETWORK_DIM": 16, "NETWORK_ALPHA": 16, "LEARNING_RATE": 1e-4,
-        "MAX_TRAIN_EPOCHS": 55, "SAVE_EVERY_N_EPOCHS": 1, "SEED": 42,
-        "ADAPTIVE_LR": True, "ADAPTIVE_LR_MIN": "1e-4", "ADAPTIVE_LR_MAX": "4e-4",
-        "TARGET_LAYERS": "Full Model", "MIN_TIMESTEP": "", "MAX_TIMESTEP": "",
-        "OPTIMIZER_TYPE": "adamw8bit",
-    },
-    "✨ Old Reliable - Flavour 8 (rank 8, full model, single subject)": {
-        "NETWORK_DIM": 8, "NETWORK_ALPHA": 8, "LEARNING_RATE": 1e-4,
-        "MAX_TRAIN_EPOCHS": 55, "SAVE_EVERY_N_EPOCHS": 1, "SEED": 42,
-        "ADAPTIVE_LR": True, "ADAPTIVE_LR_MIN": "1e-4", "ADAPTIVE_LR_MAX": "4e-4",
-        "TARGET_LAYERS": "Full Model", "MIN_TIMESTEP": "", "MAX_TIMESTEP": "",
-        "OPTIMIZER_TYPE": "adamw8bit",
-    },
-    "✨ Identity (rank 4, single subject)": {
-        "NETWORK_DIM": 4, "NETWORK_ALPHA": 4, "LEARNING_RATE": 4e-4,
-        "MAX_TRAIN_EPOCHS": 15, "SAVE_EVERY_N_EPOCHS": 1, "SEED": 42,
-        "ADAPTIVE_LR": True, "ADAPTIVE_LR_MIN": "2e-4", "ADAPTIVE_LR_MAX": "4e-4",
-        "TARGET_LAYERS": "Identity", "MIN_TIMESTEP": "", "MAX_TIMESTEP": "",
-        "OPTIMIZER_TYPE": "adamw8bit",
-    },
-    "✨ Identity (rank 8, harder dataset)": {
-        "NETWORK_DIM": 8, "NETWORK_ALPHA": 8, "LEARNING_RATE": 4e-4,
-        "MAX_TRAIN_EPOCHS": 20, "SAVE_EVERY_N_EPOCHS": 1, "SEED": 42,
-        "ADAPTIVE_LR": True, "ADAPTIVE_LR_MIN": "2e-4", "ADAPTIVE_LR_MAX": "4e-4",
-        "TARGET_LAYERS": "Identity", "MIN_TIMESTEP": "", "MAX_TIMESTEP": "",
-        "OPTIMIZER_TYPE": "adamw8bit",
-    },
-    "✨ Multi-Character (rank 16, multi character or concept)": {
-        "NETWORK_DIM": 16, "NETWORK_ALPHA": 16, "LEARNING_RATE": 2e-4,
-        "MAX_TRAIN_EPOCHS": 50, "SAVE_EVERY_N_EPOCHS": 1, "SEED": 42,
-        "ADAPTIVE_LR": True, "ADAPTIVE_LR_MIN": "1e-4", "ADAPTIVE_LR_MAX": "4e-4",
-        "TARGET_LAYERS": "Identity", "MIN_TIMESTEP": "", "MAX_TIMESTEP": "",
-        "OPTIMIZER_TYPE": "adamw8bit",
-    },
-    "✨ Style (late timesteps)": {
-        "NETWORK_DIM": 4, "NETWORK_ALPHA": 4, "LEARNING_RATE": 4e-4,
-        "MAX_TRAIN_EPOCHS": 15, "SAVE_EVERY_N_EPOCHS": 1, "SEED": 42,
-        "ADAPTIVE_LR": True, "ADAPTIVE_LR_MIN": "1e-5", "ADAPTIVE_LR_MAX": "4e-4",
-        "TARGET_LAYERS": "Style", "MIN_TIMESTEP": "0", "MAX_TIMESTEP": "400",
-        "OPTIMIZER_TYPE": "adamw8bit",
-    },
-    "✨ Style+Composition (all timesteps)": {
-        "NETWORK_DIM": 4, "NETWORK_ALPHA": 4, "LEARNING_RATE": 4e-4,
-        "MAX_TRAIN_EPOCHS": 15, "SAVE_EVERY_N_EPOCHS": 1, "SEED": 42,
-        "ADAPTIVE_LR": True, "ADAPTIVE_LR_MIN": "1e-5", "ADAPTIVE_LR_MAX": "4e-4",
-        "TARGET_LAYERS": "Style+Composition", "MIN_TIMESTEP": "", "MAX_TIMESTEP": "",
-        "OPTIMIZER_TYPE": "adamw8bit",
-    },
-}
-
-# Built-in presets for Krea 2. The Klein block-targeting / adaptive-LR / timestep presets
-# above don't apply (Krea 2 trains all linears, no block map yet, no adaptive LR), so Krea 2
-# ships a single sensible-defaults entry. Users can still save their own via Save Preset —
-# those land in the per-architecture preset folder and appear alongside this one.
-KREA2_BUILT_IN_PRESETS = {
-    "✨ Krea 2 Defaults (rank 32, full model)": {
-        "NETWORK_DIM": 32, "NETWORK_ALPHA": 32, "NETWORK_TYPE": "LoRA (standard)",
-        "LEARNING_RATE": 1e-4,
-        "MAX_TRAIN_EPOCHS": 30, "SAVE_EVERY_N_EPOCHS": 1, "SEED": 42,
-        "ADAPTIVE_LR": False, "ADAPTIVE_LR_MIN": "1e-4", "ADAPTIVE_LR_MAX": "4e-4",
-        "TARGET_LAYERS": "Full Model", "MIN_TIMESTEP": "", "MAX_TIMESTEP": "",
-        "OPTIMIZER_TYPE": "adamw8bit",
-        "GRADIENT_ACCUMULATION": 1, "MAX_GRAD_NORM": 1.0,
-        "DATASET_MEGAPIXELS": "0.25",
-        # Memory settings all auto — each resolves from the actual GPU at launch.
-        # BLOCKS_SWAP must be the combobox's exact label: _apply_preset_values matches a
-        # preset value against the offered options on its first token, case-sensitively,
-        # so a bare "auto" would not select "Auto (detect from GPU)".
-        "BLOCKS_SWAP": "Auto (detect from GPU)",
-        "QUANT_4BIT_MODE": "auto", "COMPILE_BLOCKS": "Auto",
-        # Per-image loss watch: detection + the LR throttle on, the two interventions that
-        # rewrite captions or pre-judge images left off — those want a deliberate choice.
-        "KREA2_LOSS_WATCH": True, "KREA2_PER_IMAGE_LR": True,
-        "KREA2_AUTO_RECAPTION": False, "KREA2_WARMUP_LOOK": False,
-    },
-    # Rank 8 + Adaptive LR at an aggressive floor: fewer epochs to a usable LoRA. Everything
-    # else identical to Krea 2 Defaults (which stays the preset applied on family switch).
-    "✨ Krea 2 Ultra Fast (rank 8, adaptive LR)": {
-        "NETWORK_DIM": 8, "NETWORK_ALPHA": 8, "NETWORK_TYPE": "LoRA (standard)",
-        "LEARNING_RATE": 1e-4,
-        "MAX_TRAIN_EPOCHS": 20, "SAVE_EVERY_N_EPOCHS": 1, "SEED": 42,
-        "ADAPTIVE_LR": True, "ADAPTIVE_LR_MIN": "2e-4", "ADAPTIVE_LR_MAX": "4e-4",
-        "TARGET_LAYERS": "Full Model", "MIN_TIMESTEP": "", "MAX_TIMESTEP": "",
-        "OPTIMIZER_TYPE": "adamw8bit",
-        "GRADIENT_ACCUMULATION": 1, "MAX_GRAD_NORM": 1.0,
-        "DATASET_MEGAPIXELS": "0.25",
-        "BLOCKS_SWAP": "Auto (detect from GPU)",
-        "QUANT_4BIT_MODE": "auto", "COMPILE_BLOCKS": "Auto",
-        "KREA2_LOSS_WATCH": True, "KREA2_PER_IMAGE_LR": True,
-        "KREA2_AUTO_RECAPTION": False, "KREA2_WARMUP_LOOK": False,
-    },
-    # Style. Two deliberate departures from the identity-shaped presets above:
-    #
-    #   Rank 16 — style is a broad global direction (palette, texture, rendering), not the
-    #   fine detail identity memorises. The only style-specific parameter advice published
-    #   for Krea 2 is about rank, not LR ("rank 16 for simple styles, 64 for complex
-    #   subjects" — Krea2Trainer), so 16 follows the one datapoint that exists.
-    #
-    #   LR ceiling 2e-4, not 4e-4 — this matters MORE than the floor. The adaptive watcher
-    #   probes UP x1.25 on steady descent, and style loss descends steadily early, so it
-    #   climbs toward whatever ceiling it's given; 4e-4 is where style overbake lives. The
-    #   geometric-midpoint start rule puts min 5e-5 / max 2e-4 at exactly 1e-4 — the LR the
-    #   whole Krea 2 ecosystem defaults to (Krea's own docs, Krea2Trainer, RunComfy) — while
-    #   the watcher keeps its protection in both directions.
-    #
-    # Fewer epochs (15) because style overbakes fast, and on Krea 2 that shows up as
-    # generations dragging toward the training set's COMPOSITIONS, not just its look —
-    # so save every epoch and scrub for the sweet spot in LoRA Royale.
-    "✨ Krea 2 Style (rank 16, gentle LR)": {
-        "NETWORK_DIM": 16, "NETWORK_ALPHA": 16, "NETWORK_TYPE": "LoRA (standard)",
-        "LEARNING_RATE": 1e-4,
-        "MAX_TRAIN_EPOCHS": 15, "SAVE_EVERY_N_EPOCHS": 1, "SEED": 42,
-        "ADAPTIVE_LR": True, "ADAPTIVE_LR_MIN": "5e-5", "ADAPTIVE_LR_MAX": "2e-4",
-        "TARGET_LAYERS": "Full Model", "MIN_TIMESTEP": "", "MAX_TIMESTEP": "",
-        "OPTIMIZER_TYPE": "adamw8bit",
-        "GRADIENT_ACCUMULATION": 1, "MAX_GRAD_NORM": 1.0,
-        "DATASET_MEGAPIXELS": "0.25",
-        "BLOCKS_SWAP": "Auto (detect from GPU)",
-        "QUANT_4BIT_MODE": "auto", "COMPILE_BLOCKS": "Auto",
-        # Detection + throttle on as usual. Look-outlier warm-up stays off for an extra
-        # reason here: it scores images by FACE embedding, which is meaningless on a style
-        # set that may have no faces at all.
-        "KREA2_LOSS_WATCH": True, "KREA2_PER_IMAGE_LR": True,
-        "KREA2_AUTO_RECAPTION": False, "KREA2_WARMUP_LOOK": False,
-    },
-}
-
-# LoRA Royale seed-travel presets — recipes of the *mechanics* knobs (reference /
-# anchor / journey / identity-lock), NOT a prompt system (seed travel uses the Setup
-# prompt). Reference is kept light (0.1–0.4) so the seeds can actually travel.
-SEED_TRAVEL_PRESETS = {
-    # Each varies a clearly different knob:
-    "Hold subject":   dict(ref_strength="0.4",  ref_mp="1.0", sequential=False, waypoints="2"),
-    "Journey":        dict(ref_strength="0.2",  ref_mp="1.0", sequential=False, waypoints="6"),
-    "Feedback dream": dict(ref_strength="0.25", ref_mp="1.0", sequential=True,  waypoints="4"),
-    # No reference at all — pure noise wandering through many seeds.
-    "Free ride":      dict(ref_strength="0.0",  ref_mp="1.0", sequential=False, waypoints="8"),
-}
-
-# MiniMax H3 built-in presets — barebones image-only LoRA. Only the knobs the H3 trainer reads
-# apply (rank/alpha/lr/epochs/save/seed/adaptive/optimizer/grad-accum/max-grad-norm/megapixels);
-# the H3 base is always NF4 (no swap / fp8 / quant knobs). The first entry is applied on switch.
-MINIMAX_BUILT_IN_PRESETS = {
-    # LoKR factor 8 rather than standard LoRA: the same call Krea 2 landed on after measurement
-    # (highest likeness recorded here, no skin sheen). Dim/alpha still ride along for when the
-    # Network Type is flipped back to LoRA — the H3 trainer ignores them under LoKR.
-    #
-    # Static 1e-4, adaptive OFF. The adaptive watcher probes the LR UP on steady descent, and a
-    # density sweep needs the LR held still or the runs aren't comparable.
-    #
-    # 0.5 MP rather than the 0.25 the other families default to: detail has to be present in the
-    # training image before any schedule can teach it, and 0.5 still fits without block swap.
-    #
-    # 60% low-noise WITH mid-concentrated. The two controls answer different questions: the box
-    # is how much training sits below sigma 0.5, the checkbox is where that mass actually lands.
-    # At 60% the split is 58.8% of steps in sigma 0.3-0.7, 12.9% below 0.2, 5.0% above 0.8
-    # (measured, 2M draws through sample_sigmas). The same 60% on a UNIFORM base would dump far
-    # more into the near-clean end, where the latent is almost the finished image and the step
-    # teaches little. Peter ran 70% + mid-concentrated well; 60 keeps slightly more mid-band and
-    # gives back some sigma>0.8 mass, which is the only part of the schedule that lets the LoRA
-    # touch composition at all. Both beat the 22% this shipped with.
-    #
-    # Note the mid-band fraction peaks around 50% and is nearly flat from 45-60, so it does NOT
-    # discriminate on its own — it is a description of where the steps go, not a prediction of
-    # likeness. 60 is Peter's call from real runs, with the arithmetic agreeing rather than
-    # leading. 50% is worth trying: its shift is exactly 1.00, i.e. the plain logit-normal that
-    # earlier runs recorded as overdriving adapters — on adamw8bit, before the optimizer was
-    # understood.
-    # LoRA, not LoKR (Peter, 10 Aug). The runs that produced the best likeness on this family
-    # were standard LoRA at dim/alpha 16, and LoKR moves ~7-10x further per unit LR — which made
-    # the same Learning Rate box mean two very different things depending on the Network Type
-    # sitting above it. LoKR stays one dropdown away for anyone who wants it.
-    "✨ MiniMax H3 (Lower LR - slower)": {
-        "NETWORK_DIM": 16, "NETWORK_ALPHA": 16,
-        "NETWORK_TYPE": "LoRA (standard)", "LOKR_FACTOR": 8,
-        # Flat 1e-4 (Peter, 17 Aug). With the ramp off this IS the rate — and rank 16 wants
-        # half of what the rank-8 Fast preset runs at (which keeps its flat 2e-4).
-        "LEARNING_RATE": 1e-4,
-        # Ships OFF (Peter, 17 Aug — reversing 11 Aug): the slow build spent the early epochs
-        # crawling and the flat 2e-4 runs have been the ones delivering. The ramp stays a
-        # dropdown away for anyone who wants the held-ratio start.
-        "MINIMAX_ADAPTER_RAMP": "Off",
-        # Carried so "load Defaults" genuinely resets it. Multi Concept overrides this to Off
-        # when it is on, and the command builder locks it there regardless.
-        "MINIMAX_CAPTION_DROPOUT": "0.05 (default)",
-        "MAX_TRAIN_EPOCHS": 60, "SAVE_EVERY_N_EPOCHS": 1, "SEED": 42,
-        "ADAPTIVE_LR": False, "ADAPTIVE_LR_MIN": "1e-5", "ADAPTIVE_LR_MAX": "4e-4",
-        # adamw, NOT adamw8bit — the single biggest likeness change measured on H3 (2026-08-06).
-        # Every other knob had been swept with likeness stuck around 40-50%; full-precision
-        # optimizer state moved it night-and-day on the same dataset. The 8-bit optimizer stores
-        # the second moment blockwise-quantized, and on this model that is evidently costing the
-        # fine detail. Costs ~1.2 GB of fp32 state against a 21 GB resident base.
-        "OPTIMIZER_TYPE": "adamw",
-        "GRADIENT_ACCUMULATION": 1, "MAX_GRAD_NORM": 1.0,
-        # Back to 0.25 MP (Peter, 11 Aug). The 1.0 default lasted a day: it came from the theory
-        # that 496px trains below H3's 768 short-edge canvas and must therefore starve detail —
-        # the same out-of-distribution argument that was being made about previews at the time.
-        # A day of real runs did not bear it out, and 0.25 is four times cheaper per step.
-        # The canvas number is about what the model RENDERS; it turned out to say much less than
-        # expected about what it can be TRAINED on.
-        "DATASET_MEGAPIXELS": "0.25",
-        "MINIMAX_LOWNOISE_PCT": "60", "MINIMAX_HIGHNOISE_LR_PCT": "100",
-        # The experiment knobs all ship OFF, so the preset is the plain baseline every A/B is
-        # measured against. Each of these was built to be TRIED, not to be on by default:
-        #   blocks "all"      — no block-range restriction
-        #   AdaLN False       — Peter's call from real runs; the reference trains it, we do not
-        #   slow blocks ""    — one LR everywhere, no depth split
-        #   distill False     — ordinary photo training, no r2v teacher (which also needs the
-        #                       _teref cache built, so defaulting it on would break a fresh run)
-        "MINIMAX_BLOCKS": "all", "MINIMAX_BASE_QUANT": MINIMAX_BASE_QUANT_OPTIONS[0],
-        "MINIMAX_TRAIN_ADALN": False,
-        # Optimised Likeness Learning ships ON: photos train the identity blocks (20-49) only,
-        # clips train the full model. The one measured exception is style — the Style preset
-        # turns it off (style needs the early blocks).
-        "MINIMAX_LIKENESS_OPT": True,
-        "MINIMAX_SLOW_BLOCKS": "", "MINIMAX_SLOW_LR_SCALE": "0.2",
-        # The one experiment that graduated: the limiter ships ON. Validated on a real A/B
-        # (8 Aug) — the last trained block always hogs 2-4x the median block's movement and
-        # over-edits fine detail (distorted eyes); capping it fixed epoch-1 quality outright
-        # Per-step clip and LR warmup are RETIRED (Peter, 10 Aug) and hidden — both were
-        # after-the-fact answers to epoch-1 overshoot, which the Adapter-relative LR ramp
-        # removes at its root by holding the step/size ratio steady. Kept as keys so old
-        # presets and saved configs still load; the command builder never emits them.
-        "MINIMAX_BLOCK_LIMIT": "Off",
-        "MINIMAX_LR_WARMUP": "Off",
-        # EMA stays available but OFF by default: it saves the smoothed centre of the stride
-        # zigzag instead of a raw corner of it, which is worth having when a run is pushed hard
-        # and unnecessary when it is not.
-        "MINIMAX_EMA": "Off",
-        "MINIMAX_DISTILL": False,
-    },
-    # (An earlier "Fast (adaptive LR)" preset was retired 9 Aug when its recipe drifted from the
-    # main one. The Fast preset below avoids that by DERIVING from Defaults — see the comment.)
-}
-
-# --- MiniMax H3 Fast ------------------------------------------------------------------------
-# Peter's rank-8 recipe (11 Aug): the run that reached full likeness in ~400 steps and came out
-# noticeably more flexible than the rank-16 ones. Low rank cannot memorise backgrounds and
-# framing in the time available, so it is forced to encode the subject instead.
-#
-# Built by SPREADING Defaults rather than copying it, so a change to the shipped recipe cannot
-# silently leave this one behind — which is exactly how the last Fast preset earned its
-# retirement. Keyed off the first entry rather than the title, because the title has been
-# renamed twice already.
-_MM_DEFAULTS_KEY = next(iter(MINIMAX_BUILT_IN_PRESETS))
-MINIMAX_BUILT_IN_PRESETS["✨ MiniMax H3 Fast (LoRA 8, 50 epochs)"] = {
-    **MINIMAX_BUILT_IN_PRESETS[_MM_DEFAULTS_KEY],
-    "NETWORK_DIM": 8, "NETWORK_ALPHA": 8,
-    "MAX_TRAIN_EPOCHS": 50,
-    "LEARNING_RATE": 2e-4,
-    # Flat, not ramped. The ramp exists to stop a full-size stride landing on a near-zero
-    # adapter; at rank 8 there are half as many directions to move, and the measured run that
-    # this preset reproduces had no ramp at all.
-    "MINIMAX_ADAPTER_RAMP": "Off",
-    "ADAPTIVE_LR": False,
-}
-
-# --- MiniMax H3 Style -------------------------------------------------------------------------
-# The 19 Aug style ablation (Repair Studio, same instrument that found the likeness set): a
-# style LoRA's deltas matter across nearly the WHOLE model — droppable only at 4-5 (the dead
-# band / audio-embedder pipe) and 48-49 (subject-specific last-mile work: load-bearing for
-# likeness and voice, silent for style). Hence 0-3, 6-47. LR matches Fast's 2e-4 — Peter's
-# real style runs (20 Aug) found the halved 1e-4 unnecessary; drop it manually for an extra-
-# gentle run if a style ever fries.
-MINIMAX_BUILT_IN_PRESETS["✨ MiniMax H3 Style (LoRA 8)"] = {
-    **MINIMAX_BUILT_IN_PRESETS["✨ MiniMax H3 Fast (LoRA 8, 50 epochs)"],
-    "LEARNING_RATE": 2e-4,
-    "MINIMAX_BLOCKS": "0-3, 6-47",
-    # MUST be off here: style measurably needs the early blocks the likeness mask freezes, and
-    # with it on the blocks spec above would be ignored outright.
-    "MINIMAX_LIKENESS_OPT": False,
-}
-
-# Fast is the shipped default (Peter, 22 Aug): the FIRST entry is what a family switch and a
-# fresh start apply, and the rank-8 Fast recipe is where most datasets should begin. The
-# rank-16 recipe stays one dropdown away, flagged in the GUI as the larger-dataset choice.
-# (Re-inserting an existing key keeps its first position, so Fast leads and nothing else moves.)
-_MM_FAST_KEY = "✨ MiniMax H3 Fast (LoRA 8, 50 epochs)"
-MINIMAX_BUILT_IN_PRESETS = {
-    _MM_FAST_KEY: MINIMAX_BUILT_IN_PRESETS[_MM_FAST_KEY],
-    **MINIMAX_BUILT_IN_PRESETS,
-}
-
-# Directory for dataset configurations
-DATASET_DIR = os.path.join(os.path.dirname(__file__), "dataset")
-
-# Directory for cache (latents and text encoder outputs)
-CACHE_DIR = os.path.join(os.path.dirname(__file__), "cache")
-
-# Directory for output LoRAs
-OUTPUT_LORAS_DIR = os.path.join(os.path.dirname(__file__), "output_loras")
-
-# File for storing last-used folder paths
-LAST_USED_FILE = os.path.join(os.path.dirname(__file__), ".last_used.json")
-
-
-def load_last_used():
-    """Load last-used folder paths from config file"""
-    defaults = {
-        "image_prep_source": "",
-        "image_folder": "",  # Start tab: training image folder (shared with Captions)
-        "image_folder2": "",  # Multi Concept (MiniMax): second subject, TRAINING ONLY
-        "caption_trigger": "trigger_word",
-        "dataset_cache_dir": CACHE_DIR,
-        "sample_prompt": "A high quality photo",
-    }
-    if os.path.exists(LAST_USED_FILE):
-        try:
-            with open(LAST_USED_FILE, 'r', encoding='utf-8') as f:
-                saved = json.load(f)
-                defaults.update(saved)
-        except Exception:
-            pass
-    # Migrate pre-Start-tab last_used files: if image_folder isn't set but one
-    # of the legacy keys (caption_folder / dataset_image_dir / image_prep_source)
-    # has a value, seed image_folder from the best candidate.
-    if not defaults.get("image_folder"):
-        for legacy_key in ("caption_folder", "dataset_image_dir", "image_prep_source"):
-            legacy_val = defaults.get(legacy_key, "")
-            if legacy_val:
-                defaults["image_folder"] = legacy_val
-                break
-    return defaults
-
-
-def save_last_used(data):
-    """Save last-used folder paths to config file.
-
-    Atomic (tmp + os.replace): this file is rewritten on every traced-var edit, and its
-    reader falls back to defaults on a JSONDecodeError — so a crash mid-write used to
-    silently blank the remembered paths, and the next auto-save persisted the blanks."""
-    try:
-        tmp = LAST_USED_FILE + ".tmp"
-        with open(tmp, 'w', encoding='utf-8') as f:
-            json.dump(data, f, indent=2)
-        os.replace(tmp, LAST_USED_FILE)
-    except Exception:
-        pass
-
-
-# ---------------------------------------------------------------------------
-# Preferences — centralized model and directory paths (Klein 9B only for now)
-# ---------------------------------------------------------------------------
-
-PREFS_FILE = os.path.join(os.path.dirname(__file__), "prefs.json")
-HELP_FILE = os.path.join(os.path.dirname(__file__), "help.json")
-_FIZGIG_DIR = os.path.dirname(os.path.abspath(__file__))
-
-# --- Captioning models -----------------------------------------------------
-# Florence-2 downloads itself from HuggingFace on first use; the Qwen3-VL entry appears only when
-# the Krea 2 text-encoder file is present in Preferences. That file is a full vision-language
-# model with a real LM head, so it can genuinely describe an image — and the captions it writes
-# are plain .txt, so it is useful for ANY dataset, Klein runs included. The label names where the
-# file comes from, not who it is for.
-FLORENCE_DEFAULT_MODEL = "MiaoshouAI/Florence-2-base-PromptGen"
-FLORENCE_MODELS = [FLORENCE_DEFAULT_MODEL, "microsoft/Florence-2-base", "microsoft/Florence-2-large"]
-# Florence-2 isn't a native transformers architecture, so loading it means trust_remote_code=True
-# — downloading and EXECUTING whatever Python is currently on that repo's default branch, with no
-# pin. Pinned here to the commit each was audited against, so a compromised account (or a repo
-# that just changes later) can't silently change what gets executed on someone's next first-run.
-# To refresh a pin: check https://huggingface.co/api/models/<repo> for the current "sha".
-FLORENCE_REVISIONS = {
-    "MiaoshouAI/Florence-2-base-PromptGen": "da7ac9f3deac56a928e2fd4d94d8bb985d231299",
-    "microsoft/Florence-2-base": "5ca5edf5bd017b9919c05d08aebef5e4c7ac3bac",
-    "microsoft/Florence-2-large": "21a599d414c4d928c9032694c424fb94458e3594",
-}
-# PromptGen's config doesn't carry its own modeling code — its auto_map points at
-# "microsoft/Florence-2-base-ft--modeling_florence2...", so transformers fetches the code that
-# actually EXECUTES from that second, different repo. Our `revision` above only pins PromptGen
-# itself; transformers only carries it over to the code repo automatically when the two repos
-# are the same one (they aren't here), so the redirected repo needs its own explicit pin via
-# code_revision or it silently stays on "main". The two microsoft/ models don't redirect
-# (their auto_map has no repo prefix, just the module path), so they don't need an entry here.
-FLORENCE_CODE_REVISIONS = {
-    "MiaoshouAI/Florence-2-base-PromptGen": "f6c1a25888ffc1d945ee8a1a77ac833c7303d46e",  # microsoft/Florence-2-base-ft
-}
-FLORENCE_TASKS = ["<CAPTION>", "<DETAILED_CAPTION>", "<MORE_DETAILED_CAPTION>"]
-QWEN_CAPTION_MODEL = "Qwen3-VL 4B (Krea 2 text encoder)"
-QWEN_CUSTOM_TASK = "Custom…"
-
-# Prefs whose VALUE is a directory we want to keep portable across repo
-# clones/moves. When saved to disk, paths inside _FIZGIG_DIR are stored as
-# relative strings (with forward slashes); when loaded, they're resolved back
-# to absolute so every consumer works unchanged. Paths outside the repo (e.g.
-# a user pointing Cache to another drive) stay absolute on both sides.
-#
-# Model paths (base_dit, vae, text_encoder, ...) are NOT in this set — they
-# point to external HuggingFace downloads and are always absolute.
-#
-# Note: `dataset_dir` is deliberately NOT a pref — the dataset TOML path is
-# fully hardcoded to FIZGIG_DIR/dataset/ via the DATASET_DIR module constant
-# (Dataset tab auto-saves to Fizgig_train.toml; no browse UI anywhere), so
-# exposing it in Preferences was dead weight.
-_PORTABLE_DIR_KEYS = {"lora_output_dir", "profiles_dir", "cache_dir"}
-
-
-def _resolve_pref_path(value: str) -> str:
-    """Convert a possibly-relative stored path to an absolute path by joining
-    with _FIZGIG_DIR. Absolute paths are returned unchanged."""
-    if not value:
-        return value
-    if os.path.isabs(value):
-        return value
-    return os.path.normpath(os.path.join(_FIZGIG_DIR, value))
-
-
-def _serialize_pref_path(value: str) -> str:
-    """Store a path as relative-to-_FIZGIG_DIR if it lives inside the repo;
-    otherwise store as an absolute path. Uses forward slashes for
-    cross-platform portability of the JSON file."""
-    if not value:
-        return value
-    try:
-        abs_value = os.path.abspath(value)
-        fizgig_abs = os.path.abspath(_FIZGIG_DIR)
-        rel = os.path.relpath(abs_value, fizgig_abs)
-    except ValueError:
-        # Windows: path is on a different drive than the repo — relpath raises.
-        return os.path.abspath(value).replace(os.sep, "/")
-    if rel.startswith("..") or os.path.isabs(rel):
-        # Outside the repo — keep absolute.
-        return os.path.abspath(value).replace(os.sep, "/")
-    return rel.replace(os.sep, "/")
-
-
-DEFAULT_PREFS = {
-    # Model paths (absolute — point to external model downloads).
-    # Blank on first launch; user fills these in via the Preferences tab. Each
-    # row has a "Download" link that opens the correct HuggingFace repo.
-    "base_dit": "",
-    "distilled_dit": "",
-    "vae": "",
-    "text_encoder": "",
-    # Krea 2 model paths. RAW = training base; Turbo (pre-quant fp8) = previews/
-    # inference; Qwen-Image VAE; Qwen3-VL-4B text encoder (bf16 for training).
-    "krea2_raw_dit": "",
-    "krea2_turbo_dit": "",
-    "krea2_vae": "",
-    "krea2_text_encoder": "",
-    # Turbo distillation LoRA (rank 64) — RAW + this at strength 1.0 behaves as the Turbo
-    # model, so samples can render on the resident training DiT instead of loading the
-    # separate Turbo checkpoint (saves the park-to-CPU shuffle during previews).
-    "krea2_turbo_lora": "",
-    # MiniMax H3 model paths (experimental third family — barebones image-only LoRA training).
-    # bf16 DiT is the training base (NF4-quantized at load); Qwen3-VL-32B TE + video VAE cache.
-    "minimax_dit": "",
-    # ref2va is a DIFFERENT fine-tune from fl2va, not another quantization of it: it is what
-    # ComfyUI's r2v workflow loads, and the only H3 build that accepts reference images.
-    # Optional — required only for reference distillation.
-    "minimax_ref_dit": "",
-    "minimax_text_encoder": "",
-    "minimax_vae": "",
-    # Audio VAE — optional, and only video clips ever use it. With it, a clip's sound becomes a
-    # real training target; without it, clips train video only, which is what every dataset did
-    # before clips existed. Never loaded for a stills folder.
-    "minimax_audio_vae": "",
-    # Turbo LoRA — optional, previews only: 6-step in-training samples with the community Turbo
-    # applied at ~75% on top of the training adapter, exactly how fast ComfyUI inference runs it.
-    "minimax_turbo_lora": "",
-    # Output directories — relative to repo root, portable across clones/moves.
-    # Resolved to absolute in load_prefs(); in-memory pref values are absolute.
-    # All three live as top-level folders inside the repo:
-    #   FizgigIndependent/output_loras/
-    #   FizgigIndependent/profiles/
-    #   FizgigIndependent/cache/
-    # (Dataset TOMLs are always in FIZGIG_DIR/dataset/ via DATASET_DIR — not
-    # a pref; see _PORTABLE_DIR_KEYS note above.)
-    "lora_output_dir": "output_loras",
-    "profiles_dir": "profiles",
-    "cache_dir": "cache",
-    # Input directories — default folders the Browse dialogs open in, so users
-    # don't re-hunt every session. input_lora_dir seeds LoRA pickers (Repair
-    # Studio primary/donor, LoRA the Explorer, Context LoRA on Training);
-    # input_ref_dir seeds the reference-image pickers (Repair Studio, Explorer).
-    # Absolute paths (may live anywhere); empty = no default (last folder).
-    "input_lora_dir": "",
-    "input_ref_dir": "",
-    # Where the Start tab's training-folder Browse opens. On a pod the entrypoint seeds this to
-    # /workspace/datasets so Browse lands where the uploads are, rather than at cwd.
-    "input_dataset_dir": "",
-    # Stop the RunPod pod when a training run finishes cleanly. Off by default: a rented GPU bills
-    # by the hour, so a run that ends at 4am otherwise bills until someone notices — but stopping
-    # a machine out from under someone has to be something they asked for.
-    "runpod_stop_when_done": "0",
-    # An account API key that can stop pods. Stored here rather than expected as a template env
-    # var because a PUBLIC template hands its variables to everyone who deploys it — one person's
-    # key would end up controlling their account from strangers' containers. Kept on the user's
-    # own volume, entered in the RunPod card, masked in the UI.
-    "runpod_api_key": "",
-    # Inference DiT block swap — int 0-16 for Klein 9B. With the Distilled fp8
-    # model (workbench default) 0 = no swap fits ~16GB; loading Base is heavier
-    # (0 ≈ 24GB). 16 = max swap for the smallest cards. Applies to Repair Studio,
-    # Profiler, and Extractor. The Training tab has its own separate BLOCKS_SWAP.
-    "inference_blocks_to_swap": "Auto (detect from GPU)",
-    # INT8 fast inference: quantize the workbench/preview DiT's block Linears to int8 (W8A8) for a
-    # faster matmul. On by default — same VRAM as fp8 (8-bit either way), composes with block swap,
-    # and only affects previews (never the saved LoRA). Toggle off in Preferences to use fp8.
-    "inference_int8": "1",
-    # Which physical GPU to use, as a bare index ("0", "1", ...). Empty = leave the machine's
-    # default alone, which is what every install before this had, so single-GPU users see no
-    # change. Applied by exporting CUDA_VISIBLE_DEVICES (see _apply_cuda_device_pref): torch
-    # then renumbers the chosen card to cuda:0 and nothing downstream - trainer, loader,
-    # sampler, cache scripts - has to know a choice was made at all.
-    "cuda_device": "",
-}
-
-
-def _enumerate_gpus():
-    """[(index, name, total_gb, uuid)] for every card in the machine, or [] if it cannot be read.
-
-    Deliberately NOT torch.cuda: touching it creates the CUDA context, which fixes the visible
-    device set for the life of the process - i.e. asking torch what GPUs exist would defeat the
-    setting this list feeds. NVML and nvidia-smi both enumerate the real hardware regardless of
-    CUDA_VISIBLE_DEVICES, which is exactly what a chooser needs. UUID is included so that
-    CUDA_VISIBLE_DEVICES can be set by UUID — immune to NVML vs. CUDA index reordering on
-    Windows when the display GPU is not the fastest card (issue #104)."""
-    try:
-        import pynvml
-        pynvml.nvmlInit()
-        out = []
-        for i in range(pynvml.nvmlDeviceGetCount()):
-            h = pynvml.nvmlDeviceGetHandleByIndex(i)
-            name = pynvml.nvmlDeviceGetName(h)
-            uuid = pynvml.nvmlDeviceGetUUID(h)
-            uuid = uuid.decode() if isinstance(uuid, bytes) else uuid
-            out.append((i, name.decode() if isinstance(name, bytes) else name,
-                        pynvml.nvmlDeviceGetMemoryInfo(h).total / (1024 ** 3), uuid))
-        if out:
-            return out
-    except Exception:
-        pass
-    try:
-        import subprocess
-        r = subprocess.run(["nvidia-smi", "--query-gpu=index,name,memory.total,uuid",
-                            "--format=csv,noheader,nounits"],
-                           capture_output=True, text=True, timeout=6,
-                           creationflags=getattr(subprocess, "CREATE_NO_WINDOW", 0))
-        out = []
-        for line in r.stdout.strip().splitlines():
-            idx, name, mb, uuid = [p.strip() for p in line.split(",")]
-            out.append((int(idx), name, int(mb) / 1024, uuid))
-        return out
-    except Exception:
-        return []
-
-
-def _apply_cuda_device_pref(prefs) -> str:
-    """Export CUDA_VISIBLE_DEVICES from the saved pref. Returns what was applied, or "".
-
-    This is the whole GPU-selection mechanism: everything downstream keeps asking for "cuda"
-    and gets the chosen card, because it is the only one it can see. Must run before anything
-    creates a CUDA context - the variable is read once, when that context is built.
-
-    An existing CUDA_VISIBLE_DEVICES in the environment wins: someone who launched with it set
-    meant it, and silently overriding that from a saved pref would be worse than not having the
-    pref at all."""
-    if os.environ.get("CUDA_VISIBLE_DEVICES"):
-        return os.environ["CUDA_VISIBLE_DEVICES"]
-    want = str(prefs.get("cuda_device", "")).strip()
-    if want:
-        os.environ["CUDA_VISIBLE_DEVICES"] = want
-        return want
-    return ""
-
-
-def _auto_detect_blocks_to_swap() -> int:
-    """Pick a DiT block-swap preset based on the GPU's total VRAM.
-
-    Only called when the user hasn't saved an explicit preference yet.
-    Returns the leading integer for the swap-preset labels (0/4/8/12/16).
-    """
-    try:
-        import torch
-        if torch.cuda.is_available():
-            vram_gb = torch.cuda.get_device_properties(0).total_memory / (1024 ** 3)
-            # 16 GB and up → no swap. The Distilled fp8 model (workbench default) is
-            # light, so 16 GB cards run previews without swapping, and skipping swap
-            # avoids the PCIe latency. Threshold is 15, not 16: a 16 GB card reports
-            # ~15.9 GiB total (drivers reserve a little), so a >=16 gate would wrongly
-            # exclude it. Only genuinely smaller cards (<15 GB) swap.
-            if vram_gb >= 15:
-                return 0   # 16 GB / 24 GB / 32 GB — no swap needed
-            if vram_gb >= 10:
-                return 12  # 12-14 GB
-            return 16      # <10 GB — maximum swap
-    except Exception:
-        pass
-    return 0  # safe fallback
-
-
-def load_prefs() -> dict:
-    """Load user preferences from prefs.json, falling back to defaults.
-    Relative portable-dir paths are resolved to absolute for in-memory use."""
-    prefs = dict(DEFAULT_PREFS)
-    user_set_swap = False
-    if os.path.exists(PREFS_FILE):
-        try:
-            with open(PREFS_FILE, 'r', encoding='utf-8') as f:
-                saved = json.load(f)
-                user_set_swap = "inference_blocks_to_swap" in saved
-                prefs.update(saved)
-        except Exception:
-            pass
-    # Inference block swap defaults to "Auto (detect from GPU)", resolved at each
-    # pipeline load (see _get_inference_blocks_to_swap) — same behaviour as the
-    # training Blocks Swap setting. Users who saved an explicit value keep it.
-    first_run = not os.path.exists(PREFS_FILE)
-    _ = user_set_swap  # retained for clarity; no longer forces a concrete value
-    # Resolve portable directory paths to absolute.
-    for key in _PORTABLE_DIR_KEYS:
-        if key in prefs and isinstance(prefs[key], str):
-            prefs[key] = _resolve_pref_path(prefs[key])
-    # On first run, persist defaults so prefs.json exists immediately.
-    if first_run:
-        save_prefs(prefs)
-    return prefs
-
-
-def _persist_disabled() -> bool:
-    """True when FIZGIG_NO_PERSIST is set — headless test harnesses set it so instantiating the
-    GUI and poking vars can NEVER overwrite the user's real prefs.json / last_used.json /
-    settings / Fizgig_train.toml (traced vars auto-save on write, so a test setting
-    image_folder_var would otherwise clobber the remembered training folder)."""
-    return bool(os.environ.get("FIZGIG_NO_PERSIST"))
-
-
-def _running_on_pod() -> bool:
-    """True when this is the Docker pod image (RunPod, or any rented box).
-
-    Keyed off OUR marker, set in docker/entrypoint.sh, rather than RunPod's RUNPOD_POD_ID: the
-    hosting provider's variable names are theirs to change, and another host would set different
-    ones entirely. The pod id below is read separately and only used for display and for targeting
-    a stop — everything degrades if it is absent."""
-    return os.environ.get("FIZGIG_POD", "0") not in ("0", "", None)
-
-
-def _pod_id() -> str:
-    """The provider's id for this pod, or "" if it doesn't advertise one."""
-    for key in ("RUNPOD_POD_ID", "HOSTNAME"):
-        v = (os.environ.get(key) or "").strip()
-        if v:
-            return v
-    return ""
-
-
-def _app_commit() -> str:
-    """Short git commit of the running checkout, or "".
-
-    On a pod this is the interesting half of the version: the IMAGE is pinned in the RunPod
-    template while the app pulls master at every boot, so the two are meant to differ and a bug
-    report needs both."""
-    try:
-        import subprocess as _sp
-        r = _sp.run(["git", "-C", FIZGIG_DIR, "rev-parse", "--short", "HEAD"],
-                    capture_output=True, text=True, timeout=8,
-                    creationflags=(0x08000000 if os.name == "nt" else 0))
-        return r.stdout.strip() if r.returncode == 0 else ""
-    except Exception:
-        return ""
-
-
-def _git(*args, timeout=8) -> str:
-    """Run a git command in the repo, return stripped stdout or "" on any failure."""
-    try:
-        import subprocess as _sp
-        r = _sp.run(["git", "-C", FIZGIG_DIR, *args],
-                    capture_output=True, text=True, timeout=timeout,
-                    creationflags=(0x08000000 if os.name == "nt" else 0))
-        return r.stdout.strip() if r.returncode == 0 else ""
-    except Exception:
-        return ""
-
-
-def _git_describe_version() -> str:
-    """Human version of the running checkout: 'v3.1.1' exactly on a tag,
-    'v3.1.1-2-gee3a7fa' in between. Pods clone --depth 1 (no tags), so describe falls back
-    to a bare short SHA — which read as a mystery build number in the field ('Version
-    37c0c2f' was current master, taken for an old app). Name the branch so a tagless
-    checkout says what it is: 'master @ 37c0c2f'."""
-    v = _git("describe", "--tags", "--always") or _app_commit()
-    if v and "v" not in v.split("-")[0]:
-        branch = _git("rev-parse", "--abbrev-ref", "HEAD")
-        if branch and branch != "HEAD":
-            return f"{branch} @ {v}"
-    return v
-
-
-def _latest_release_tag():
-    """Highest vX.Y.Z release tag on origin as (tag, sha), or None when unreachable.
-
-    ls-remote is a single read-only round-trip — no fetch, no GitHub API, no tokens or
-    rate limits. Returning None (offline, no origin, ZIP download) means 'unknown':
-    the caller shows nothing rather than a false nag."""
-    out = _git("ls-remote", "--tags", "--refs", "origin", "v*", timeout=15)
-    if not out:
-        return None
-    best, best_key = None, None
-    for line in out.splitlines():
-        try:
-            sha, ref = line.split(None, 1)
-            tag = ref.strip().rsplit("/", 1)[-1]
-            key = tuple(int(p) for p in tag.lstrip("v").split("."))
-        except (ValueError, IndexError):
-            continue    # not a plain vX.Y.Z tag — ignore
-        if best_key is None or key > best_key:
-            best, best_key = (tag, sha), key
-    return best
-
-
-def _update_status_from(latest, has_obj: bool, is_ancestor: bool) -> str:
-    """Pure decision: 'up_to_date' | 'update_available' | 'unknown'.
-
-    Up to date means HEAD CONTAINS the latest release tag's commit — true both exactly on
-    the tag and ahead of it (a user who pulled newer master must not be nagged). A tag whose
-    commit we don't even have locally, or have but isn't an ancestor, means the release is
-    newer than this checkout."""
-    if latest is None:
-        return "unknown"
-    if has_obj and is_ancestor:
-        return "up_to_date"
-    return "update_available"
-
-
-def _git_ok(*args, timeout=8) -> bool:
-    """Run git for its EXIT CODE (merge-base --is-ancestor answers that way)."""
-    try:
-        import subprocess as _sp
-        r = _sp.run(["git", "-C", FIZGIG_DIR, *args],
-                    capture_output=True, text=True, timeout=timeout,
-                    creationflags=(0x08000000 if os.name == "nt" else 0))
-        return r.returncode == 0
-    except Exception:
-        return False
-
-
-def _check_for_update():
-    """Full check (network + local git). Returns ('update_available', tag) /
-    ('up_to_date', current) / ('unknown', '')."""
-    latest = _latest_release_tag()
-    if latest is None:
-        return "unknown", ""
-    tag, sha = latest
-    has_obj = _git_ok("cat-file", "-e", f"{sha}^{{commit}}")
-    is_anc = has_obj and _git_ok("merge-base", "--is-ancestor", sha, "HEAD")
-    status = _update_status_from(latest, has_obj, is_anc)
-    if status == "update_available":
-        # Shallow-clone truth (pods clone --depth 1): the tag's commit isn't in the local
-        # object store even when HEAD is AHEAD of it, so ancestry can't clear us — but being
-        # exactly the remote tip can. Without this, every pod born after any post-release
-        # commit showed a FALSE Update Available banner on perfectly current code (the 19 Aug
-        # "old pods" saga — three images, all current, all nagging). A genuinely stale pod
-        # is neither on the tag nor at the tip, so real updates still flag.
-        remote_head = (_git("ls-remote", "origin", "HEAD", timeout=15) or "").split()
-        local_head = _git("rev-parse", "HEAD")
-        if remote_head and local_head and remote_head[0] == local_head:
-            status = "up_to_date"
-    return status, (tag if status == "update_available" else _git_describe_version())
-
-
-def _pod_stop_key_env() -> str:
-    """A stop-capable key from the environment, or "".
-
-    NOT RunPod's injected RUNPOD_API_KEY — that one is pod-scoped and 403s on every pod-management
-    call (verified on a live pod, and a documented RunPod limitation). Stopping a pod needs an
-    account key, which is why this reads a separate variable rather than falling back."""
-    return (os.environ.get("RUNPOD_STOP_API_KEY") or "").strip()
-
-
-def save_prefs(prefs: dict) -> None:
-    """Save preferences to prefs.json. Portable-dir paths inside the repo are
-    stored as relative strings so a cloned/moved repo finds its own defaults."""
-    if _persist_disabled():
-        return
-    to_save = {}
-    for key, value in prefs.items():
-        if key in _PORTABLE_DIR_KEYS and isinstance(value, str):
-            to_save[key] = _serialize_pref_path(value)
-        else:
-            to_save[key] = value
-    try:
-        # Atomic: a truncated prefs.json silently blanked every model path (the reader
-        # swallows JSONDecodeError and falls back to defaults, then the next auto-save
-        # persisted the blanks).
-        tmp = PREFS_FILE + ".tmp"
-        with open(tmp, 'w', encoding='utf-8') as f:
-            json.dump(to_save, f, indent=2)
-        os.replace(tmp, PREFS_FILE)
-    except Exception:
-        pass
-
-
-# Maps Training tab settings keys to prefs keys — entries with matching keys
-# will be bound to the shared prefs StringVar (two-way sync with Preferences tab)
-SETTING_TO_PREF = {
-    "DIT_MODEL": "base_dit",
-    "VAE_MODEL": "vae",
-    "TEXT_ENCODER": "text_encoder",
-    "LORA_OUTPUT_DIR": "lora_output_dir",
-}
-
-
-# Default preset for Klein Base 9B — the only supported architecture.
-# Applied on first launch and when "Reset to Defaults" is used.
-PRESETS = {
-    "Flux 2 Klein Base 9B": {
-        "LEARNING_RATE": 0.0004,       # Rank 4:4 optimal LR for 80+ image datasets
-        "NETWORK_DIM": 4,              # Low rank — identity signal only
-        "NETWORK_ALPHA": 4,            # 1:1 alpha:rank ratio
-        "MAX_TRAIN_EPOCHS": 12,        # Fast convergence at low rank
-        "OPTIMIZER_TYPE": "adamw8bit",
-        "TIMESTEP_SAMPLING": "flux2_shift",
-        "DISCRETE_FLOW_SHIFT": "0",
-        "WEIGHTING_SCHEME": "none",
-        "BLOCKS_SWAP": "auto",
-        # Model paths come from Preferences at runtime — leave blank in the preset.
-        "VAE_MODEL": "",
-        "TEXT_ENCODER": "",
-        "DIT_MODEL": "",
-        "FP8": True,
-        "SCALED": True,  # BF16 model, use fp8_scaled for memory efficiency
-        "QUANT_4BIT": False,  # 4-bit NF4 base (low-VRAM); supersedes fp8 when on
-        "COMPILE_BLOCKS": "auto",  # torch.compile the DiT blocks (krea2): auto | on | off
-        "GRADIENT_CHECKPOINTING": True,  # ON by default — needed to fit 9B on most cards
-    },
-}
-
+from fizgig_gui.core.domain.minimax_math import (
+    MINIMAX_LOWNOISE_SIGMA, minimax_lownoise_to_shift, minimax_highnoise_lr,
+    minimax_shift_to_lownoise, MINIMAX_STRUCTURE_OPTIONS, MINIMAX_STRUCTURE_DESC,
+    MINIMAX_STRUCTURE_DEFAULT, MINIMAX_BLOCK_OPTIONS, MINIMAX_NUM_BLOCKS,
+    MINIMAX_LIKENESS_BLOCKS, MINIMAX_AUDIO_BLOCKS, MINIMAX_BASE_QUANT_OPTIONS,
+    minimax_base_quant, minimax_block_spec, MINIMAX_TRAIN_BASE_OPTIONS, minimax_train_base,
+)
+
+from fizgig_gui.core.config.presets import (
+    ft_checkpoint_continuation, BUILT_IN_PRESETS, KREA2_BUILT_IN_PRESETS,
+    SEED_TRAVEL_PRESETS, MINIMAX_BUILT_IN_PRESETS, _MM_DEFAULTS_KEY,
+)
+
+from fizgig_gui.core.config.last_used import (
+    DATASET_DIR, CACHE_DIR, OUTPUT_LORAS_DIR, LAST_USED_FILE,
+    load_last_used, save_last_used,
+)
+
+
+from fizgig_gui.core.config.prefs import (
+    PREFS_FILE, HELP_FILE, _FIZGIG_DIR,
+    FLORENCE_DEFAULT_MODEL, FLORENCE_MODELS, FLORENCE_REVISIONS,
+    FLORENCE_CODE_REVISIONS, FLORENCE_TASKS, QWEN_CAPTION_MODEL, QWEN_CUSTOM_TASK,
+    _PORTABLE_DIR_KEYS, _resolve_pref_path, _serialize_pref_path, DEFAULT_PREFS,
+    _enumerate_gpus, _apply_cuda_device_pref, _auto_detect_blocks_to_swap,
+    load_prefs, save_prefs,
+    _persist_disabled, _running_on_pod, _pod_id, _app_commit,
+    _git, _git_describe_version, _latest_release_tag, _update_status_from,
+    _git_ok, _check_for_update, _pod_stop_key_env,
+)
+
+
+from fizgig_gui.core.config.settings_map import SETTING_TO_PREF, PRESETS
 
 class LoRATrainerGUI:
     def __init__(self, master):
